requirements.dev.txt
Flask==1.1.2
uwsgi==2.0.18
flask_cors==3.0.7
flask_jwt_extended==3.24.1
Flask-PyMongo==2.3.0
marshmallow==3.6.0
pillow==7.1.2
requests==2.23.0
mongomock==3.19.0
coverage==5.1